refactor: extract UITestRuntimeSupport.swift from AppBootstrap.swift
Extracts the DEBUG-only `UITestRuntimeSupport` enum (with nested
`InMemory{Keychain,SessionStorage,ScreenCaptureCapability,
ScreenshotCapture,SpaceQuery,URLOpener,…}` helpers) from
AppBootstrap.swift into its own file guarded by the same `#if DEBUG`.

Byte-identical extraction of lines 78-497 from AppBootstrap.swift
(the entire `#if DEBUG ... #endif` block).

New file adds only `import Foundation` (all API references resolve to
Foundation types: URL, UUID, Date; other types are project-local).

Reduces AppBootstrap.swift from 497 to 76 lines (production surface
only); new file is 422 lines and stays out of release builds.

Tests: 667 (unchanged).
